added delete functionality, improved refresh functionality, added switch/case display functionality to QCD

+++ client/src/components/DeleteQuizButton.js
@@ -3,7 +3,10 @@ import axios from 'axios';
 
 export default class DeleteQuizButton extends Component {
     removeQuiz = () => {
-        
+        if (this.props.currentQuiz === this.props.deleteKey){
+            console.log("Cannot delete active quiz");
+            return;
+        }
         axios.delete('http://localhost:3001/quiz/deleteQuiz', {
             data: {
                 id: this.props.deleteKey,
